refactor(cats): Moving non-redis cats classes to cats-core
