+++ cats/cats-core/src/main/java/com/netflix/spinnaker/cats/cluster/DefaultNodeStatusProvider.java
@@ -14,8 +14,11 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.cats.redis.cluster;
+package com.netflix.spinnaker.cats.cluster;
 
-public interface NodeStatusProvider {
-  boolean isNodeEnabled();
+public class DefaultNodeStatusProvider implements NodeStatusProvider {
+  @Override
+  public boolean isNodeEnabled() {
+    return true;
+  }
 }
