+++ cats/cats-core/src/main/java/com/netflix/spinnaker/cats/cluster/NodeIdentity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.cats.redis.cluster;
+package com.netflix.spinnaker.cats.cluster;
 
 public interface NodeIdentity {
     String getNodeIdentity();
