+++ cats/cats-dynomite/src/main/java/com/netflix/spinnaker/cats/dynomite/cluster/DynoClusteredAgentScheduler.java
@@ -23,9 +23,9 @@ import com.netflix.spinnaker.cats.agent.AgentScheduler;
 import com.netflix.spinnaker.cats.agent.AgentSchedulerAware;
 import com.netflix.spinnaker.cats.agent.ExecutionInstrumentation;
 import com.netflix.spinnaker.cats.module.CatsModuleAware;
-import com.netflix.spinnaker.cats.redis.cluster.AgentIntervalProvider;
-import com.netflix.spinnaker.cats.redis.cluster.NodeIdentity;
-import com.netflix.spinnaker.cats.redis.cluster.NodeStatusProvider;
+import com.netflix.spinnaker.cats.cluster.AgentIntervalProvider;
+import com.netflix.spinnaker.cats.cluster.NodeIdentity;
+import com.netflix.spinnaker.cats.cluster.NodeStatusProvider;
 import com.netflix.spinnaker.cats.thread.NamedThreadFactory;
 import com.netflix.spinnaker.kork.dynomite.DynomiteClientDelegate;
 import net.jodah.failsafe.Failsafe;
