+++ cats/cats-dynomite/src/main/java/com/netflix/spinnaker/cats/dynomite/cluster/DynoClusteredSortAgentScheduler.java
@@ -20,9 +20,9 @@ import com.netflix.spinnaker.cats.agent.*;
 import com.netflix.spinnaker.cats.dynomite.DynomiteUtils;
 import com.netflix.spinnaker.cats.dynomite.ExcessiveDynoFailureRetries;
 import com.netflix.spinnaker.cats.module.CatsModuleAware;
-import com.netflix.spinnaker.cats.redis.cluster.AgentIntervalProvider;
+import com.netflix.spinnaker.cats.cluster.AgentIntervalProvider;
 import com.netflix.spinnaker.cats.redis.cluster.ClusteredSortAgentLock;
-import com.netflix.spinnaker.cats.redis.cluster.NodeStatusProvider;
+import com.netflix.spinnaker.cats.cluster.NodeStatusProvider;
 import com.netflix.spinnaker.cats.thread.NamedThreadFactory;
 import com.netflix.spinnaker.kork.jedis.RedisClientDelegate;
 import net.jodah.failsafe.Failsafe;
